+++ plugins/analysis/conflict_merge.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # This file is part of Diggit.
 #
 # Diggit is free software: you can redistribute it and/or modify
@@ -25,7 +27,7 @@ require 'yaml'
 class ConflictMerge < Diggit::Analysis
 	require_addons 'out'
 
-	DIFF3 = 'diff3'.freeze
+	DIFF3 = 'diff3'
 
 	def initialize(options)
 		super(options)