+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Version 3.0.4
 * New graph analysis
 * New diff extractor analysis
+* Updated dependencies
 
 ### Version 3.0.3
 * New yard analysis